% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_geo.R
\name{deal_geo}
\alias{deal_geo}
\title{Pipeline of processing GEO data}
\usage{
deal_geo(
  GEO_name = GEO_name,
  GPL_number = GPL_number,
  dir = dir,
  quiet = TRUE,
  rawdata = FALSE,
  Type = "AFFY",
  multiprobe = "median",
  removeRawdata = TRUE
)
}
\arguments{
\item{GEO_name}{GSE number}

\item{GPL_number}{Probe annotation platform}

\item{dir}{Des dir}

\item{quiet}{logical vector, default is FALSE}

\item{rawdata}{logical vector, default is FALSE}

\item{Type}{Type one of "AFFY", "AGI4-44K" when \code{rawdata=TRUE}}

\item{multiprobe}{multiprobe one of "median", "mean", "max", "min", default is median}

\item{removeRawdata}{logical vector, default is TRUE}
}
\value{

}
\description{
Pipeline of processing GEO data
}
\examples{
library(Dbio)
# Download matrixs and preprocess
deal_geo("GSE54566", "GPL570", "./")
# Download, preprocess matrixs and keep rawdata
deal_geo("GSE54566", "GPL570", "./", removeRawdata = FALSE)
# Download, preprocess rawdata of affymetrixs
deal_geo("GSE54566", "GPL570", "./", rawdata = TRUE, Type = "AFFY")
}
