% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_sample.R
\name{deal_sample}
\alias{deal_sample}
\title{Download, Preprocess, and Save sample metadata}
\usage{
deal_sample(GEO_name = GEO_name, dir = dir)
}
\arguments{
\item{GEO_name}{GSE number}

\item{dir}{Des dir}
}
\value{

}
\description{
Download, Preprocess, and Save sample metadata
}
\examples{
library(Dbio)
deal_sample(GEO_name = "GSE54566", dir = "./")
}
