% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_log2.R
\name{detect_log2}
\alias{detect_log2}
\title{Confirm expmat log2(+1) transfered by detection}
\usage{
detect_log2(data = expmat)
}
\arguments{
\item{data}{expression matrixs}
}
\value{
log2(+1) transformed matrixs
}
\description{
Confirm expmat log2(+1) transfered by detection
}
\examples{
library(Dbio)
expmat <- get_matrixs("GSE54566", "./")
expmat2 <- detect_log2(expmat)
}
