% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matrixs.R
\name{get_matrixs}
\alias{get_matrixs}
\title{Get matrixs data from GEO}
\usage{
get_matrixs(GEO_name = GEO_name, dir = dir)
}
\arguments{
\item{GEO_name}{GSE Number}

\item{dir}{Dest dir}
}
\value{

}
\description{
Get matrixs data from GEO
}
\examples{
library(Dbio)
get_matrixs("GSE54566", "./")
}
