% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probe2symbol.R
\name{get_probe2symbol}
\alias{get_probe2symbol}
\title{Get probe annation data}
\usage{
get_probe2symbol(
  GPL_number,
  GEO_name = GEO_name,
  dir = dir,
  Array_type = "None"
)
}
\arguments{
\item{GPL_number}{Probe annotation platfrom number}

\item{GEO_name}{GSE number}

\item{dir}{Des dir}
}
\value{

}
\description{
Get probe annation data
}
\examples{
library(Dbio)
get_probe2symbol("GPL570", "GSE54566", "./")
}
