% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample.R
\name{get_sample}
\alias{get_sample}
\title{Get clinical data from GEO}
\usage{
get_sample(GEO_name = GEO_name, dir = dir)
}
\arguments{
\item{GEO_name}{GEO number}

\item{dir}{Dest dir}
}
\value{
clinical data of GSE you choosed
}
\description{
Get clinical data from GEO
}
\examples{
#' library(Dbioinfo)
get_sample("GSE555", "./")
}
