% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_file.R
\name{initial_file}
\alias{initial_file}
\title{Initial file structure}
\usage{
initial_file(GEO_name = GEO_name, dir = dir)
}
\arguments{
\item{GEO_name}{GSE number}

\item{dir}{Dest dir}
}
\value{
GSE download File structer
}
\description{
Initial file structure
}
\examples{
library(Dbioinfo)
initial_file("GSE555", "./")
}
