% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_sample.R
\name{output_sample}
\alias{output_sample}
\title{Output clinical data from GEO to des dir}
\usage{
output_sample(sample, sample_post, GEO_name = GEO_name, dir = dir)
}
\arguments{
\item{sample}{Sample metadata get from get_sample function}

\item{sample_post}{Preprocessed sample metadata from preprocess_sample}

\item{GEO_name}{GSE Number}

\item{dir}{Des dir}
}
\value{

}
\description{
Output clinical data from GEO to des dir
}
\examples{
library(Dbio)
sample <- get_sample("GSE54566", "./")
sample_post <- preprocess_sample(sample)
output_sample(sample, sample_post, "GSE54566", "./")
}
