% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_matrixs.R
\name{preprocess_matrixs}
\alias{preprocess_matrixs}
\title{Process matrixs from GEO}
\usage{
preprocess_matrixs(
  dataset = expmat,
  metadata = sample_post,
  annotation = probe2gene,
  method.probe = "median"
)
}
\arguments{
\item{dataset}{Expression matrixs}

\item{metadata}{sample metadata}

\item{annotation}{probe annotaion file}

\item{method.probe}{one of "medain","mean","max","min", default is "median"}
}
\value{
preprocessed expression matrixs
}
\description{
Process matrixs from GEO
}
\examples{
library(Dbio)
# make file structure
initial_file("GSE54566", "./")
# get sample data
sample <- get_sample("GSE54566", "./")
# process sample
sample_post <- preprocess_sample(sample)
# get expression data
expmat <- get_matrixs("GSE54566", "./")
# get probe annotation data
annotation <- get_probe2symbol("GPL570", "GSE54566", "./")
preprocess_matrixs(
  dataset = expmat,
  metadata = sample_post,
  annotation = annotation
)
}
