% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_rawdata.R
\name{preprocess_rawdata}
\alias{preprocess_rawdata}
\title{Process raw data from GEO}
\usage{
preprocess_rawdata(
  GEO_name = GEO_name,
  annotation = probe2gene,
  arrayType = Type,
  method.probe = "median",
  dir = dir
)
}
\arguments{
\item{GEO_name}{GSE number}

\item{annotation}{Probe anotation platform}

\item{arrayType}{one of "AFFY", "AGI4-44K"}

\item{method.probe}{one of "medain", "mean", "max", "min", default is "median"}

\item{dir}{Des dir}
}
\value{

}
\description{
Process raw data from GEO
}
\examples{
library(Dbio)
# make file structure
initial_file("GSE54566", "./")
# get probe annotation data
annotation <- get_probe2symbol("GPL570", "GSE54566", "./")
preprocess_rawdata(
  GEO_name = "GSE54566",
  annotation = annotation,
  arrayType = "AFFY",
  method.probe = "median",
  dir = "./"
)
}
