% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sample.R
\name{preprocess_sample}
\alias{preprocess_sample}
\title{Preproces clinical data from GEO}
\usage{
preprocess_sample(metadata = sample)
}
\arguments{
\item{metadata}{Sample metadata get from GEO}
}
\value{

}
\description{
Preproces clinical data from GEO
}
\examples{
library(Dbio)
library(GEOquery)
sample <- get_sample("GSE54566", "./")
sample_post <- preprocess_sample(sample)
}
