% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe2multisymbol.R
\name{probe2multisymbol}
\alias{probe2multisymbol}
\title{Deal with Probe to Multi Symbol}
\usage{
probe2multisymbol(data = exp, method.probe = "median")
}
\arguments{
\item{data}{expression matrixs}

\item{method.probe}{one of "medain","mean","max","min", default is "median"}
}
\value{

}
\description{
Calculate the unique symbol (expression) matrixs
The gene symbol column should name : "gene"
}
\examples{
library(Dbio)
# get median value of same gene symbol in matrix
exp1 <- probe2multisymbol(exp)
# get max  value of same gene symbol in matrix
exp2 <- probe2multisymbol(exp, "max")
}
